feat: split out style snapshots update image used in storybook
